Rework notation system.
Each notation is declared and exported in `src/notations/` to be imported in `Juggle.notations.js`. The declaration has to define:
- `parse( string )` converts a string to `.throws` array of a siteswap,
- `unparse( throws )` converts that array into a string (which when parsed has to return the same result),
- `hands( degree )` returns the hand labels, and optionally
- `limits` object showing the restrictions of numerical properties the notation can encode

For example, the following notation, which can only describe one handed juggling, would accept nothing but the string "ok", which represents the 7 ball cascade:

{
   limits: {
      degree: { min: 1, max: 1 }
   },
   hands: function(degree){ return ["hand"]; },
   parse: function(string){ if( string !== "ok" ) throw new Error("Not ok."); return [[[{ value: 7, handFrom: 0, handTo: 0 }]]]; }
   unparse: function(throws){ return "ok"; }
}

Calling `.toString()` for an incompatible notation (determined by `.limits`) throws an error with the appropriate description.

----------

Parse grammars of all notations are defined in `src/notations/parser/grammar.ne`.

Whitespace rules are now grammar specific.

----------

Multiple notations can be passed to the `Siteswap(string, notation)` constructor as an array. The first one that parses successfully will be assigned, or "Invalid syntax." error will be thrown.

The existing notations, `standard` and `compressed` are split into `standard:async`, `standard:sync`, `compressed:async` and `compressed:sync`, but they can still be used as aliases for the pair of notations they represent, that is "compressed" is converted to ["compressed:async", "compressed:sync"].