Add multihand notation.
This one is superior to the others as any number of hands can be used.

Syntaxwise - rows represent hands and throw values are paired with a string (A,B,C,...,Z,AA,AB...) referring to the target hand. I plan to update the readme with more details for all notations.